requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
ultralytics==8.3.26
torch>=2.1.0
Pillow==10.4.0
numpy==2.1.1

